refactor: starts separating out path handling util files
